refactor: change the tests data generation
